pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "lxml>=6.0.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "plotly>=6.5.1",
    "psycopg2-binary>=2.9.11",
    "requests>=2.32.5",
    "streamlit>=1.52.2",
    "streamlit-local-storage>=0.0.25",
]
